Removed file that shouldn't be tracked + fix conversion Double in CGfloat

+++ Pharus/View/CircleView.swift
@@ -26,7 +26,7 @@ class CircleView: UIView {
     
     lazy var completionPercentage = 0.0 {
         didSet {
-            makeCircularPath(completionPercentage: completionPercentage)
+            makeCircularPath(completionPercentage: CGFloat(completionPercentage))
         }
     }
